vectordbtest/requirements.txt
aiohttp>=3.8.0
asyncio>=3.4.3
python-dotenv>=0.19.0
qdrant-client==1.7.0
numpy>=1.24.0
openai==1.12.0 